Stat_tools imports organized
