+++ gtime/plotting/plotting.py
@@ -3,7 +3,7 @@ import pandas as pd
 import numpy as np
 from typing import Union, List, Callable, Optional
 from gtime.plotting.preprocessing import seasonal_split
-from gtime.stat_tools.tools import acf, pacf
+from gtime.stat_tools import acf, pacf
 from scipy.stats import norm
