+++ gtime/stat_tools/__init__.py
@@ -1,0 +1,12 @@
+"""
+The :mod:`gtime.stat_tools` module contains statistical functions: autocorrelations, MLE estimates etc.
+"""
+
+from .mle_estimate import MLEModel
+from .tools import acf, pacf
+
+__all__ = [
+    "MLEModel",
+    "acf",
+    "pacf"
+]
